+++ zodiaclove/server/config.env
@@ -1,0 +1,1 @@
+ATLAS_URI = mongodb+srv://zodiacAdmin:4bwMKcmbZSRUi6r@cluster0.zd46e.mongodb.net/zodiac?retryWrites=true&w=majority
