+++ zodiaclove/server/database/connection.js
@@ -1,10 +1,6 @@
 const { MongoClient } = require("mongodb");
-const uri =
-  "mongodb+srv://zodiacAdmin:<4bwMKcmbZSRUi6r>@cluster0.zd46e.mongodb.net/myFirstDatabase?retryWrites=true&w=majority";
-
-const DB = process.env.uri;
-
-const client = new MongoClient(uri, {
+const Db = process.env.ATLAS_URI;
+const client = new MongoClient(Db, {
   useNewUrlParser: true,
   useUnifiedTopology: true,
 });
@@ -13,16 +9,17 @@ var _db;
 
 module.exports = {
   connectToServer: function (callback) {
-    client.connect(function (err, database) {
-      if (database) {
-        _db = database.db("myFirstDatabase");
-        console.log(`Succesful connection to mongoDB`);
+    client.connect(function (err, db) {
+      // Verify we got a good "db" object
+      if (db) {
+        _db = db.db("myFirstDatabase");
+        console.log("Successfully connected to MongoDB.");
       }
       return callback(err);
     });
   },
 
-  getDatabase: function () {
+  getDb: function () {
     return _db;
   },
 };
