+++ zodiaclove/server/server.js
@@ -1,12 +1,13 @@
 const express = require("express");
 const app = express();
 const cors = require("cors");
-//require("dotenv").config({ path: "./config.env" });
+const routes = require("./api/users");
+require("dotenv").config({ path: "./config.env" });
 const port = process.env.PORT || 3000;
 
 app.use(cors());
 app.use(express.json());
-app.use(require("./api/endpoints"));
+app.use(routes);
 
 const dbo = require("./database/connection");
